refactor(action): move `action` to separate file
